requirements.txt
pandas>=1.3.0
numpy>=1.20.0
matplotlib>=3.4.0
# yfinance>=0.2.0  # 한국투자증권 API로 대체
pykrx>=1.0.0
ta>=0.10.0
requests>=2.25.0
streamlit>=1.20.0
plotly>=5.10.0
scikit-learn>=1.0.0
python-dotenv>=0.19.0